feat(ws-17): refactor core planning to core/planning structure
WS-17 (Core Planning):
- Move planner.py → core/planning/planner.py
- Move archive.py → core/planning/archive.py
- Create core/planning/__init__.py
- Create shims in src/pipeline/ for backward compatibility
- Update core/ shims to reference core.planning.*

Tests:
- Validated: import from core.planning.planner works

Risk: MEDIUM (not heavily used - stubs/minimal implementation)
All Phase D workstreams (WS-15, WS-16, WS-17) now complete!
