+++ core/planning/planner.py
@@ -1,0 +1,56 @@
+"""
+This module provides a draft automated planner for generating workstream bundles.
+For now, it contains function skeletons and docstrings, serving as a placeholder
+for future v2.0 automation.
+"""
+
+from typing import Any, Dict, List, Optional
+
+def plan_workstreams_from_spec(spec_source: Any, options: Optional[Dict[str, Any]] = None) -> List[Dict[str, Any]]:
+    """
+    Generates a list of draft workstream bundles from a given specification source.
+
+    This function is a placeholder for future automated planning logic.
+    It will eventually:
+    - Read OpenSpec/CCPM-like inputs.
+    - Apply decomposition rules (e.g., from config/decomposition_rules.yaml).
+    - Infer file scopes based on changes or other heuristics.
+    - Potentially leverage AI tools (like Aider) to propose workstreams.
+
+    For this phase (PH-05.5), it primarily serves as a stub.
+
+    Args:
+        spec_source: The source of the specification (e.g., a path, ID, or structured dict).
+                     This is a placeholder and its exact type/content will be defined later.
+        options: Optional dictionary of planning options (e.g., max workstreams, grouping strategy).
+
+    Returns:
+        A list of dictionaries, where each dictionary represents a draft workstream bundle.
+        These bundles are intended for human review and further refinement.
+
+    Raises:
+        NotImplementedError: This function is currently a stub and does not implement
+                             full planning logic.
+    """
+    print(f"Planner v0 stub: Received spec_source={spec_source}, options={options}")
+    print("Automated planning logic is not yet implemented. See src/pipeline/planner.py.")
+    # Example of a minimal placeholder bundle
+    return [{
+        "id": "ws-generated-placeholder",
+        "openspec_change": "OS-GENERATED-001",
+        "ccpm_issue": 999,
+        "gate": 1,
+        "files_scope": ["src/placeholder.py"],
+        "tasks": ["Implement generated placeholder logic."],
+        "tool": "aider",
+        "metadata": {"notes": "This is a placeholder generated by the planner stub."}
+    }]
+    # raise NotImplementedError("Automated planning logic is not yet implemented.")
+
+if __name__ == "__main__":
+    # Example usage for the stub
+    print("Running planner.py as a script (stub mode).")
+    draft_bundles = plan_workstreams_from_spec("example_spec_id_123", {"group_by": "language"})
+    print("\nGenerated Draft Bundles (stub output):")
+    for bundle in draft_bundles:
+        print(json.dumps(bundle, indent=2))
